# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Council Asset & Compliance Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List jobs with optional filters
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 25 }
        - name: status
          in: query
          schema: { type: string }
        - name: priority
          in: query
          schema: { type: string }
        - name: assetId
          in: query
          schema: { type: string }
        - name: assignedUserId
          in: query
          schema: { type: string }
        - name: assignedTeamId
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Job list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobList"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a new job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
  /jobs/{jobId}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job detail with comments and documents
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update job fields
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Soft delete a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
  /jobs/{jobId}/complete:
    post:
      operationId: completeJob
      tags: [jobs]
      summary: Mark job as completed
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                notes: { type: string }
      responses:
        "200":
          description: Job completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{jobId}/cancel:
    post:
      operationId: cancelJob
      tags: [jobs]
      summary: Cancel a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Job cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{jobId}/follow-on:
    post:
      operationId: createFollowOnJob
      tags: [jobs]
      summary: Create a follow-on job from an existing job
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Follow-on job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{jobId}/comments:
    post:
      operationId: addJobComment
      tags: [jobs]
      summary: Add a comment to a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string }
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobComment"

  # ── SERVICE RECORDS ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Job:
      type: object
      required: [id, tenantId, title, priority, status, createdAt]
      properties:
        id: { type: string }
        tenantId: { type: string }
        title: { type: string }
        description: { type: ["string", "null"] }
        assetId: { type: ["string", "null"] }
        complianceItemId: { type: ["string", "null"] }
        assignedUserId: { type: ["string", "null"] }
        assignedTeamId: { type: ["string", "null"] }
        priority: { type: string, enum: [low, medium, high, critical] }
        dueDate: { type: ["string", "null"] }
        status: { type: string, enum: [open, assigned, in_progress, awaiting_evidence, completed, cancelled] }
        completionDate: { type: ["string", "null"] }
        completionNotes: { type: ["string", "null"] }
        parentJobId: { type: ["string", "null"] }
        createdBy: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
        asset: { type: ["object", "null"] }
        assignee: { type: ["object", "null"] }
    JobDetail:
      allOf:
        - $ref: "#/components/schemas/Job"
        - type: object
          properties:
            comments:
              type: array
              items:
                $ref: "#/components/schemas/JobComment"
            documents:
              type: array
              items:
                $ref: "#/components/schemas/Document"
            followOnJobs:
              type: array
              items:
                $ref: "#/components/schemas/Job"
    JobList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    JobInput:
      type: object
      required: [title]
      properties:
        title: { type: string }
        description: { type: string }
        assetId: { type: string }
        complianceItemId: { type: string }
        assignedUserId: { type: string }
        assignedTeamId: { type: string }
        priority: { type: string, enum: [low, medium, high, critical] }
        dueDate: { type: string }
        parentJobId: { type: string }
    JobUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        assignedUserId: { type: string }
        assignedTeamId: { type: string }
        priority: { type: string }
        dueDate: { type: string }
        status: { type: string }
        completionNotes: { type: string }
    JobComment:
      type: object
      required: [id, jobId, userId, body, createdAt]
      properties:
        id: { type: string }
        jobId: { type: string }
        tenantId: { type: string }
        userId: { type: string }
        body: { type: string }
        createdAt: { type: string }
        user: { type: ["object", "null"] }
...
    Document:
      type: object
      required: [id, tenantId, fileName, filePath, fileType, uploadedBy, createdAt]
      properties:
        id: { type: string }
        tenantId: { type: string }
        assetId: { type: ["string", "null"] }
        complianceItemId: { type: ["string", "null"] }
        jobId: { type: ["string", "null"] }
        serviceRecordId: { type: ["string", "null"] }
        fileName: { type: string }
        filePath: { type: string }
        fileType: { type: string }
        fileSize: { type: ["integer", "null"] }
        uploadedBy: { type: string }
        createdAt: { type: string }
        deletedAt: { type: ["string", "null"] }